fixup store-preset-list to not wrap on first bank and dec event, but rather stay at the current selection. had to change indicies to signed int instead of size_t for sanitize to work, closes #3069

+++ projects/epc/playground/src/proxies/hwui/panel-unit/boled/preset-screens/controls/PresetListSelectStorePosition.cpp
@@ -127,8 +127,8 @@ void PresetListSelectStorePosition::sanitizeBankPosition(PresetManager *pm)
 {
   if(auto numBanks = pm->getNumBanks())
   {
-    m_storeModeData->bankPos = std::min(m_storeModeData->bankPos, numBanks - 1);
-    m_storeModeData->bankPos = std::max(m_storeModeData->bankPos, size_t(0));
+    m_storeModeData->bankPos = std::min(m_storeModeData->bankPos, int(numBanks) - 1);
+    m_storeModeData->bankPos = std::max(m_storeModeData->bankPos, 0);
     return;
   }
 
@@ -141,8 +141,8 @@ void PresetListSelectStorePosition::sanitizePresetPosition(Bank *bank)
   {
     if(auto numPresets = bank->getNumPresets())
     {
-      m_storeModeData->presetPos = std::min(m_storeModeData->presetPos, numPresets - 1);
-      m_storeModeData->presetPos = std::max(m_storeModeData->presetPos, size_t(0));
+      m_storeModeData->presetPos = std::min(m_storeModeData->presetPos, int(numPresets) - 1);
+      m_storeModeData->presetPos = std::max(m_storeModeData->presetPos, 0);
       return;
     }
   }
